+++ Source/SmartHotel.Clients/SmartHotel.Clients.iOS/SmartHotel.Clients.iOS.csproj
@@ -23,53 +23,35 @@
   <ItemGroup>
     <ProjectReference Include="..\SmartHotel.Clients\SmartHotel.Clients.csproj" />
   </ItemGroup>
+
   <ItemGroup>
-    <PackageReference Include="Acr.UserDialogs">
-      <Version>8.1.0-alpha-0009</Version>
-    </PackageReference>
-    <PackageReference Include="Autofac">
-      <Version>4.9.2</Version>
-    </PackageReference>
-    <PackageReference Include="CarouselView.FormsPlugin">
-      <Version>6.0.0</Version>
-    </PackageReference>
-    <PackageReference Include="Microcharts">
-      <Version>0.7.1</Version>
-    </PackageReference>
-    <PackageReference Include="Microcharts.Forms">
-      <Version>0.7.1</Version>
-    </PackageReference>
-    <PackageReference Include="Microsoft.CSharp" Version="4.7.0" />
-    
-    <PackageReference Include="Refractored.MvvmHelpers">
-      <Version>1.3.0</Version>
-    </PackageReference>
-    <PackageReference Include="Rg.Plugins.Popup">
-      <Version>1.1.5.188</Version>
-    </PackageReference>
-    <PackageReference Include="SkiaSharp">
-      <Version>1.68.0</Version>
-    </PackageReference>
-    <PackageReference Include="Xamanimation">
-      <Version>1.2.0</Version>
-    </PackageReference>
-    <PackageReference Include="Xamarin.FFImageLoading">
-      <Version>2.4.5.880-pre</Version>
-    </PackageReference>
-    <PackageReference Include="Xamarin.FFImageLoading.Forms">
-      <Version>2.4.5.880-pre</Version>
-    </PackageReference>
-    
-    
+    <PackageReference Include="Acr.UserDialogs" Version="8.1.0-alpha-0009" />
+    <PackageReference Include="Autofac" Version="4.9.2" />
+
+    <PackageReference Include="Microcharts.Maui" Version="1.0.0" />
+    <PackageReference Include="Microsoft.Identity.Client" Version="4.55.0" />
+    <PackageReference Include="Refractored.MvvmHelpers" Version="1.6.2" />
+    <PackageReference Include="FFImageLoadingCompat.Maui" Version="0.1.1" />
+
+    <PackageReference Include="AlohaKit.Animations" Version="1.0.0" />
+
     <PackageReference Include="Newtonsoft.Json" Version="13.0.3" />
-    <PackageReference Include="SkiaSharp.Views.Maui.Core" Version="2.88.3" />
-    <PackageReference Include="SkiaSharp.Views.Maui.Controls" Version="2.88.3" />
-    <PackageReference Include="SkiaSharp.Views.Maui.Controls.Compatibility" Version="2.88.3" />
+    <PackageReference Include="Microsoft.Maui.Controls.Maps" Version="8.0.0-preview.7.8814" />
     <PackageReference Include="CommunityToolkit.Maui" Version="5.2.0" />
+
+    <PackageReference Include="Microsoft.Maui.Controls" Version="$(MauiVersion)" />
+    <PackageReference Include="Microsoft.Maui.Controls.Compatibility" Version="$(MauiVersion)" />
+    <PackageReference Include="Microsoft.Extensions.Logging.Debug" Version="8.0.0-preview.6.23329.7" />
+
+    <PackageReference Include="Microsoft.AppCenter" Version="5.0.2" />
+    <PackageReference Include="Microsoft.AppCenter.Analytics" Version="5.0.2" />
+    <PackageReference Include="Microsoft.AppCenter.Crashes" Version="5.0.2" />
+    <PackageReference Include="Microsoft.AppCenter.Distribute" Version="5.0.2" />
+    <PackageReference Include="Mopups" Version="1.1.1" />
   </ItemGroup>
+
   <ItemGroup>
-    <Compile Remove="Renderers\CustomNavigationRenderer.cs" />
-    <Compile Remove="Renderers\CustomPageRenderer.cs" />
+    <Folder Include="Properties\" />
   </ItemGroup>
 </Project>
